backend/api-app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
fastparquet==2023.10.1
joblib==1.3.2
pymongo[srv]==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
